runtime: move streams to new unstable namespace
