+++ components/runtime/src/ts/docs_index.ts
@@ -8,7 +8,6 @@ export * from './script';
 export * from './storage';
 export * from './models';
 export * from './httpclient';
-export * from './streams';
 export * from './scheduled_tasks';
 export * as Discord from './discord';
-
+export * as Unstable from './unstable/index';
